requirements.txt
# pysnowflake~=0.1.3
toml~=0.10.2
pysnowflake~=0.1.3
fastapi~=0.95.2
starlette~=0.27.0
pydantic~=1.10.8
pandas~=1.5.3